feat: bare-bones scouting queue
